Persistence/Basic_python_persistence/ex-1-6/pyproject.toml
[tool.poetry]
name = "ex-1-6"
version = "0.1.0"
description = ""
authors = ["siddhanthagarwal984 <siddhanthagarwal984@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
